# Improve some performance