Prefer iteration than recursion.
Because Python doesn't optimize tail recursion.

Before:
```
In [3]: %timeit expandvars(vars_)
768 µs ± 8.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

After:
```
In [3]: %timeit expandvars(vars_)
163 µs ± 826 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```